Extract commands to own file. Lib contains abstracted logic, pcsc implements NfcTransmittor
