+++ Cargo.toml
@@ -10,6 +10,7 @@ ciborium = "0.2.0"
 serde = "1"
 serde_bytes = "0.11"
 secp256k1 = { version = "0.26.0", features = ["rand-std", "bitcoin-hashes-std"] }
+hex = "0.4.3"
 
 # optional dependencies
 pcsc = { version = "2", optional = true }
